Teacher commit

+++ src/main/java/com/ahead/school/dao/StudentTestDao.java
@@ -3,5 +3,5 @@ package com.ahead.school.dao;
 import com.ahead.school.entity.StudentTest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentTestDao extends JpaRepository<StudentTest, Integer> {
+public interface StudentTestDao extends JpaRepository<StudentTest, Long> {
 }
